gestor/requirements/base.txt
pytz==2018.4  # https://github.com/stub42/pytz
python-slugify==1.2.5  # https://github.com/un33k/python-slugify
Pillow==5.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==18.1.0  # https://github.com/hynek/argon2_cffi
redis>=2.10.5  # https://github.com/antirez/redis
celery==3.1.25  # pyup: <4.0  # https://github.com/celery/celery

# Django
# ------------------------------------------------------------------------------
django==2.0.6  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.4  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.36.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.9.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

# New Packages
django-jet==1.0.7
django-reversion==2.0.13
django-fsm==2.5.0
django-elasticsearch-dsl==0.5.0
elasticsearch-dsl==6.1.0
django-tinymce==2.7.0
django-elasticsearch-dsl-drf==0.8.2
drf-extensions==0.3.1
django-cors-headers==2.2.0
pyparsing==2.2.0
pydot==1.2.4
pydotplus==2.0.2
-e git://github.com/ssvargass/django-fsm-admin.git#egg=django-fsm-admin

